Update EventButton's getX() and getY() to return doubles instead of ints
EventButton's getX() and getY() methods are now set to return doubles. The GDK
C Library uses double type property values for a GdkEventButton's X and Y. The
rationale is that, for applications and interfaces where subpixel accuracy
is required, such as on touchscreens and tablet PCs, it is better to let the
application decide if it should round off/truncate the X and Y values.

Also updated was an example in TreeView's Javadoc.

+++ src/bindings/org/gnome/gdk/EventButton.java
@@ -18,6 +18,7 @@ package org.gnome.gdk;
  * being held down by the user with {@link #getState() getState()}.
  * 
  * @author Andrew Cowie
+ * @author Srichand Pendyala
  * @since 4.0.6
  */
 public final class EventButton extends Event
@@ -53,15 +54,8 @@ public final class EventButton extends Event
      * 
      * @since 4.0.9
      */
-    /*
-     * Why on earth are they double? No one has been able to give us a
-     * coherent answer, and all the C code examples I've seen relating to
-     * TreeView selections cast these fields to (gint) so we'll do the same.
-     * FUTURE if anyone does an Override for this, then do this in C so we
-     * save shipping the double across the boundary.
-     */
-    public int getX() {
-        return (int) GdkEventButton.getX(this);
+    public double getX() {
+        return GdkEventButton.getX(this);
     }
 
     /**
@@ -70,7 +64,7 @@ public final class EventButton extends Event
      * 
      * @since 4.0.9
      */
-    public int getY() {
-        return (int) GdkEventButton.getY(this);
+    public double getY() {
+        return GdkEventButton.getY(this);
     }
 }
